Move SetHeadingLevel -> args.rs
