+++ src/ugly_widget/mod.rs
@@ -1,2 +1,3 @@
+pub mod args;
 pub mod radio_button;
 pub mod ugly_list;
